+++ livre/lib/configuration/VolumeRendererParameters.h
@@ -38,6 +38,7 @@ struct VolumeRendererParameters : public Parameters
     uint32_t minLOD; //!< Minimum level of detail
     uint32_t maxLOD; //!< Maximum level of detail
     uint32_t samplesPerRay; //!< Number of samples per ray
+    uint32_t samplesPerPixel; //!< Number of samples per ray
     std::string transferFunction; //!< Path to transfer function file
 
     /**